@@ -1,51 +1,34 @@
 project(crave)
-cmake_minimum_required(VERSION 2.8.9)
+cmake_minimum_required(VERSION 3.16.0)
 
-set(${PROJECT_NAME}_CONFIG_DIR "share/${PROJECT_NAME}" CACHE PATH
-  "where to install CMake and Make Config files")
+set(${PROJECT_NAME}_CONFIG_DIR
+    "share/${PROJECT_NAME}"
+    CACHE PATH "where to install CMake and Make Config files")
 
-############################################################
-##                CRAVE settings                          ##
-############################################################
+# ##############################################################################
+# CRAVE settings                          ##
+# ##############################################################################
 
-option(CRAVE_ENABLE_EXPERIMENTAL "enable experimental extensions of CRAVE" on)
-option(CRAVE_ENABLE_TESTS        "build tests" on)
-option(CRAVE_BUILD_EXAMPLES      "build and test examples" on)
+option(CRAVE_ENABLE_TESTS "build tests" off)
+option(CRAVE_BUILD_EXAMPLES "build and test examples" off)
 
-### C++11 required
+# C++17 required
 include(CheckCXXCompilerFlag)
-CHECK_CXX_COMPILER_FLAG("-std=c++11" COMPILER_SUPPORTS_CXX11)
-if(COMPILER_SUPPORTS_CXX11)
-  add_definitions(-std=c++11)
+check_cxx_compiler_flag("-std=c++17" COMPILER_SUPPORTS_CXX17)
+if(COMPILER_SUPPORTS_CXX17)
+  add_definitions(-std=c++17)
 else()
-  message(FATAL_ERROR "The compiler ${CMAKE_CXX_COMPILER} does not support -std=c++11. Please use a different C++ compiler.")
+  message(
+    FATAL_ERROR
+      "The compiler ${CMAKE_CXX_COMPILER} does not support -std=c++17. Please use a different C++ compiler."
+  )
 endif()
 
-### packages
+# packages
 set(CMAKE_MODULE_PATH ${PROJECT_SOURCE_DIR}/cmake)
 
-set(Boost_NO_BOOST_CMAKE true)
-set(Boost_NO_SYSTEM_PATHS true)
-
-if (WIN32)
-  if (MSVC AND (NOT MSVC_VERSION LESS 1900))
-    find_package(Boost REQUIRED)
-  else()
-    message(FATAL_ERROR "CRAVE Windows build only supports Visual Studio C++ 2015 or newer version")
-  endif()
-else()
-  find_package(Boost REQUIRED COMPONENTS system filesystem)
-endif()
-
-find_package(GLOG)
-
-if (GLOG_FOUND)
-  add_definitions(-DCRAVE_HAVE_GLOG)
-endif()
-
-if (WITH_SYSTEMC)
-  find_package(SystemC REQUIRED)
-  find_package(Threads REQUIRED)
+if(WITH_SYSTEMC)
+  find_package(SystemCLanguage REQUIRED)
 endif()
 
 add_subdirectory(metaSMT)
