@@ -2,12 +2,7 @@
 
 #pragma once
 
-#include <boost/mpl/int.hpp>
-#include <boost/mpl/sizeof.hpp>
-#include <boost/mpl/times.hpp>
-#include <boost/static_assert.hpp>
-#include <boost/type_traits/is_signed.hpp>
-#include <boost/utility/enable_if.hpp>
+#include <type_traits>
 
 namespace crave {
 
@@ -34,16 +29,16 @@ struct vector_tag;
 struct placeholder_tag;
 
 template <typename T, typename Enable = void>
-struct bitsize_traits : boost::mpl::int_<0> {
+struct bitsize_traits : std::integral_constant<std::size_t, 0> {
   static_assert(sizeof(T) == 0, "zero-sized type");
 };
 
 template <typename Integer>
-struct bitsize_traits<Integer, typename boost::enable_if<boost::is_integral<Integer> >::type>
-    : boost::mpl::times<boost::mpl::sizeof_<Integer>, boost::mpl::int_<8> >::type {};
+struct bitsize_traits<Integer, typename std::enable_if<std::is_integral<Integer>::value>::type>
+    : std::integral_constant<std::size_t, sizeof(Integer) * 8> {};
 
 template <>
-struct bitsize_traits<bool> : boost::mpl::int_<1> {};
+struct bitsize_traits<bool> : std::integral_constant<std::size_t, 1> {};
 
 template <>
 struct bitsize_traits<placeholder_tag> : public bitsize_traits<unsigned> {};
@@ -69,29 +64,26 @@ struct bitsize_traits<read_ref_tag<T> > : public bitsize_traits<T> {};
 template <typename T>
 struct bitsize_traits<vector_tag<T> > : public bitsize_traits<T> {};
 
-template <typename Integral, Integral I>
-struct bitsize_traits<boost::mpl::integral_c<Integral, I> > : public bitsize_traits<Integral> {};
-
 template <typename Enum>
-struct bitsize_traits<Enum, typename boost::enable_if<boost::is_enum<Enum> >::type> : bitsize_traits<int>::type {};
+struct bitsize_traits<Enum, typename std::enable_if<std::is_enum<Enum>::value>::type> : bitsize_traits<int>::type {};
 
 template <typename T>
-struct is_crave_variable : boost::mpl::false_ {};
+struct is_crave_variable : std::false_type {};
 
 template <typename T>
-struct is_crave_variable<randv<T> > : boost::mpl::true_ {};
+struct is_crave_variable<randv<T> > : std::true_type {};
 
 template <typename T>
-struct is_crave_variable<write_ref_tag<T> > : boost::mpl::true_ {};
+struct is_crave_variable<write_ref_tag<T> > : std::true_type {};
 
 template <typename T>
-struct is_signed : public boost::is_signed<T> {};
+struct is_signed : public std::is_signed<T> {};
 
 /**
  * assume bool to be signed (i.e. will be converted to signed bv)
  */
 template <>
-struct is_signed<bool> : public boost::is_signed<int> {};
+struct is_signed<bool> : public std::is_signed<int> {};
 
 template <typename T>
 struct is_signed<crave::randv<T> > : public crave::is_signed<T> {};
@@ -113,6 +105,9 @@ struct is_signed<crave::vector_tag<T> > : public crave::is_signed<T> {};
 
 template <typename T, typename Enable = void>
 struct to_constant_expr {};
+
+template <typename T>
+struct is_sysc_dt : std::false_type {};
 
 }  // namespace crave
 
