@@ -5,10 +5,8 @@
 #include <map>
 #include <string>
 
-#include "Context.hpp"
 #include "UserConstraintType.hpp"
 #include "UserVectorConstraint.hpp"
-#include "visitor/FixWidthVisitor.hpp"
 
 namespace crave {
 
@@ -31,27 +29,16 @@ struct ConstraintManager {
 
   void resetChanged();
 
-  template <typename Expr>
-  ConstraintPtr makeConstraint(std::string const& name, int c_id, Expr e, Context* ctx, bool const soft = false,
-                               bool const cover = false) {
-    FixWidthVisitor fwv;
-    NodePtr n(fwv.fixWidth(*boost::proto::eval(e, *ctx)));
-    return ConstraintManager::makeConstraint(name, c_id, n, ctx, soft, cover);
-  }
-
-  ConstraintPtr makeConstraint(std::string const& name, int c_id, NodePtr n, Context* ctx, bool const soft = false,
+  ConstraintPtr makeConstraint(std::string const& name, int c_id, NodePtr n, bool const soft = false,
                                bool const cover = false);
 
-  template <typename Expr>
-  ConstraintPtr makeConstraint(std::string const& name, Expr e, Context* ctx, bool const soft = false,
-                               bool const cover = false) {
-    return makeConstraint(name, new_constraint_id(), e, ctx, soft, cover);
+  ConstraintPtr makeConstraint(std::string const& name, NodePtr e, bool const soft = false, bool const cover = false) {
+    return makeConstraint(name, new_constraint_id(), e, soft, cover);
   }
 
-  template <typename Expr>
-  ConstraintPtr makeConstraint(Expr e, Context* ctx, bool const soft = false, bool const cover = false) {
+  ConstraintPtr makeConstraint(NodePtr e, bool const soft = false, bool const cover = false) {
     int id = new_constraint_id();
-    return makeConstraint("constraint_" + std::to_string(id), id, e, ctx, soft, cover);
+    return makeConstraint("constraint_" + std::to_string(id), id, e, soft, cover);
   }
 
   std::ostream& printDotGraph(std::ostream& os);
