@@ -2,6 +2,7 @@
 
 #pragma once
 
+#include <cassert>
 #include <map>
 #include <stack>
 #include <utility>
@@ -50,7 +51,7 @@ class EvalVisitor : NodeVisitor {
   virtual void visitPlusOpr(PlusOpr const&);
   virtual void visitMinusOpr(MinusOpr const&);
   virtual void visitMultipliesOpr(MultipliesOpr const&);
-  virtual void visitDevideOpr(DevideOpr const&);
+  virtual void visitDivideOpr(DivideOpr const&);
   virtual void visitModuloOpr(ModuloOpr const&);
   virtual void visitShiftLeftOpr(ShiftLeftOpr const&);
   virtual void visitShiftRightOpr(ShiftRightOpr const&);
