@@ -3,6 +3,7 @@
 #pragma once
 
 #include <map>
+#include <memory>
 #include <stack>
 #include <vector>
 
@@ -13,7 +14,7 @@ namespace crave {
 
 class ReplaceVisitor : public NodeVisitor {
  public:
-  explicit ReplaceVisitor(std::vector<boost::intrusive_ptr<VariableExpr> > *vars)
+  explicit ReplaceVisitor(std::vector<std::shared_ptr<VariableExpr>> *vars)
       : vec_idx_(), okay_(true), result_(), aux_stack_(), subscript_stack_(), variables_(vars), terminals_() {}
 
   virtual void visitNode(Node const &);
@@ -46,7 +47,7 @@ class ReplaceVisitor : public NodeVisitor {
   virtual void visitPlusOpr(PlusOpr const &);
   virtual void visitMinusOpr(MinusOpr const &);
   virtual void visitMultipliesOpr(MultipliesOpr const &);
-  virtual void visitDevideOpr(DevideOpr const &);
+  virtual void visitDivideOpr(DivideOpr const &);
   virtual void visitModuloOpr(ModuloOpr const &);
   virtual void visitShiftLeftOpr(ShiftLeftOpr const &);
   virtual void visitShiftRightOpr(ShiftRightOpr const &);
@@ -75,7 +76,7 @@ class ReplaceVisitor : public NodeVisitor {
   NodePtr result_;
   std::stack<NodePtr> aux_stack_;
   std::stack<int> subscript_stack_;
-  std::vector<boost::intrusive_ptr<VariableExpr> > *variables_;
+  std::vector<std::shared_ptr<VariableExpr>> *variables_;
   std::map<int, NodePtr> terminals_;
 };
 
